refactor: remake formSingleInput
